backend/requirements-minimal.txt
# Minimal requirements for Render deployment
fastapi==0.100.1
uvicorn==0.22.0
pydantic==1.10.12
pydantic-settings==2.0.3
sqlalchemy==1.4.48
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0